opening transition
