+++ .idea/modules.xml
@@ -5,6 +5,7 @@
       <module fileurl="file://$PROJECT_DIR$/CS188Project3.iml" filepath="$PROJECT_DIR$/CS188Project3.iml" />
       <module fileurl="file://C:\Android\AndroidStudioProjects\CS188Project3\CS188Project3.iml" filepath="C:\Android\AndroidStudioProjects\CS188Project3\CS188Project3.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
+      <module fileurl="file://C:\Users\Matthew\Desktop\Software Engineering\Project3\CS188Project3\app\app.iml" filepath="C:\Users\Matthew\Desktop\Software Engineering\Project3\CS188Project3\app\app.iml" />
       <module fileurl="file://C:\Android\AndroidStudioProjects\CS188Project3\app\app.iml" filepath="C:\Android\AndroidStudioProjects\CS188Project3\app\app.iml" />
     </modules>
   </component>
