+++ app/src/main/res/anim/first.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<alpha
+    xmlns:android="http://schemas.android.com/apk/res/android"
+android:fromAlpha="0.0"
+android:toAlpha="1.0"
+android:duration="2000" />
